move data file
